Add files via upload
Logical Operators
Logical operators are fundamental components of programming languages that allow you to combine multiple conditions and make decisions based on the result. They are often used in conditional statements (if, while, etc.) and boolean expressions to control the flow of a program.

Types of Logical Operators:
AND (&&): Returns true if both operands are true.
OR (||): Returns true if at least one of the operands is true.
NOT (!): Negates the boolean value of its operand.

Logical Operators
Logical operators are fundamental components of programming languages that allow you to combine multiple conditions and make decisions based on the result. They are often used in conditional statements (if, while, etc.) and boolean expressions to control the flow of a program.

Types of Logical Operators:
AND (&&): Returns true if both operands are true.
OR (||): Returns true if at least one of the operands is true.
NOT (!): Negates the boolean value of its operand.